Added Unit Tests and new Exception Handlers

+++ src/main/resources/data.sql
@@ -6,9 +6,9 @@ CREATE TABLE fuel_consumption (
   fuel_consumption_id BIGINT PRIMARY KEY,
   driver_id BIGINT NOT NULL,
   fuel_type VARCHAR(64) NOT NULL,
-  price_per_liter_in_eur FLOAT NOT NULL,
-  volume_in_liters FLOAT NOT NULL,
-  total_price_in_eur FLOAT,
+  price_per_liter_in_eur DECIMAL(10,5) NOT NULL,
+  volume_in_liters DECIMAL(10,5) NOT NULL,
+  total_price_in_eur DECIMAL(10,5),
   date DATE NOT NULL
 );
